Extract StreamProcessorFactory to file and begin refactor
